[edit]開眼度測定、両目の目頭の距離で補正+顔の傾きにより補正。講義開始、終了の時刻表示

+++ db_setup.py
@@ -28,8 +28,8 @@ def init_db():
             id INTEGER PRIMARY KEY AUTOINCREMENT,
             student_id INTEGER NOT NULL,
             timestamp DATETIME,
-            right_eye_openness REAL,
-            left_eye_openness REAL,
+            right_eye_openness INTEGER,
+            left_eye_openness INTEGER,
             FOREIGN KEY (student_id) REFERENCES students(id)
         )
     ''')
